compiler: Wire `{TyAnd,}Layout` into `rustc_abi`
This finally unites TyAndLayout, Layout, and LayoutS into the same crate,
as one might imagine they would be placed. No functional changes.
